+++ src/typing_ex/handler/core.clj
@@ -19,12 +19,11 @@
    [taoensso.carmine :as car]
    [taoensso.timbre :as t]))
 
-(defn- check-env []
-  (doseq [v [:prod :stage :port :auth :postgres-password :database-url
-             :redis :tp-start]]
-    (t/info v (env v))))
-
-(check-env)
+; (defn- check-env []
+;   (doseq [v [:prod :stage :port :auth :postgres-password :database-url
+;              :redis :tp-start]]
+;     (t/info v (env v))))
+; (check-env)
 
 (defonce  my-conn-pool (car/connection-pool {}))
 (def      my-conn-spec {:uri (or (env :redis) "redis://redis:6379")})
@@ -34,6 +33,8 @@
 (def ^:private redis-expire 3600)
 
 (def typing-start (or (env :tp-start) "2025-09-01"))
+(def ^:private thres-count 30)
+(def ^:private thres-point 1000)
 
 (defn admin? [s]
   (let [admins #{"hkimura"}]
@@ -53,6 +54,36 @@
    (get-in req [:headers "x-real-ip"])
    (get req :remote-addr)))
 
+(defn- smiles [n thres]
+  ;(apply str (mapv (fn [_] "🙂") (range (quot n thres))))
+  (if  (< n thres)
+    ""
+    "🙂"))
+
+(defmethod ig/init-key :typing-ex.handler.core/weekly-points [_ {:keys [db]}]
+  (fn [request]
+    (let [login (get-login request)]
+      (view/page
+       [:div
+        [:h2 (format "Weekly Points (%s)" login)]
+        (view/headline 1)
+        [:p "1 週間ごとの練習回数とスコア。🙂の数が平常点だな。"]
+        [:table.table.table-striped
+         [:thead
+          [:tr [:th "week"] [:th "回数"]　[:th "点数"]]]
+         [:tbody
+          (for [{:keys [week count pt]} (results/weekly-points db login)]
+            [:tr [:td week]
+             [:td count (smiles count thres-count)]
+             [:td pt (smiles pt thres-point)]])]]
+        [:div
+         [:ul
+          [:li "一回の練習には 1 分しかかからない。10 回練習しても 10 分だ。"]
+          [:li "30 点はかなり低い点数。 10 回練習すれば 300 点 取れる。"]
+          [:li "一週間に 3 日練習したら、回数は 30 回、点数は 1000点 くらいになる。"]
+          [:li "30 回、1000 点を超えたら、その週のタイピング平常点は 1 。"]
+          [:li "過去週のデータは書き変わらない。"]]]]))))
+
 ;; day-by-day
 (defmethod ig/init-key :typing-ex.handler.core/day-by-day [_ {:keys [db]}]
   (fn [request]
@@ -63,14 +94,19 @@
         (let [results (results/day-by-day db login)]
           (view/page
            [:div
-            [:h2 (format "Typing: last 7 days (%s)" login)]
-            ;; defined in view/page.clj as a private function
+            [:h2 (format "Typing: Last 7 days (%s)" login)]
             (view/headline 1)
             [:br]
+            [:p "直近の7日間のタイピング練習に入った時刻とスコア。"]
             [:ol {:style "margin-left:1rem;"}
-             (for [[date pt] results]
-               [:li date " " pt])]]))))))
-
+             (println results)
+             (for [{:keys [timestamp pt]} results]
+               [:li (subs (str timestamp) 0 16) " " pt])]]))))))
+(comment
+  (def s (java.util.Date.))
+  s
+  (str s)
+  (jt/format "YYYY-MM-DD HH:mm" s))
 ;; exam!
 (defmethod ig/init-key :typing-ex.handler.core/exam! [_ _]
   (fn [{{:keys [login count pt]} :params}]
@@ -126,7 +162,7 @@
 (defmethod ig/init-key :typing-ex.handler.core/login-post [_ _]
   (fn [{[_ {:strs [login password]}] :ataraxy/result}]
     (if (and (seq login) (auth? login password))
-      (-> (redirect "/total/7")
+      (-> (redirect "/todays")
           (assoc-in [:session :identity] (keyword login)))
       (-> (redirect "/login")
           (dissoc :session)
@@ -191,17 +227,11 @@
           addr (str (remote-ip req))]
       (t/info (str "/typing " user " from " addr))
       (if (roll-call-time?)
-        (cond (local? addr)
-              (typing-ex req)
-              (vpn? addr)
-              [::response/ok
-               "出席（前半）のタイプは VPN 不可。"]
-              (not (tobata? addr))
-              [::response/ok
-               "出席（前半）のタイプは学外からはできない。"]
-              :else
-              (typing-ex req))
-        (typing-ex req)))))
+        (cond (local? addr) (typing-ex req);
+              (vpn? addr) [::response/ok "出席（前半）のタイプは VPN 不可。"]
+              (not (tobata? addr)) [::response/ok "出席（前半）のタイプは学外からはできない。"]
+              :else  (typing-ex req));
+        (typing-ex req)))));
 
 (defmethod ig/init-key :typing-ex.handler.core/total [_ {:keys [db]}]
   (fn [{[_ n] :ataraxy/result :as req}]